create elodyne_packet.hpp
